app.js refactor / code splitting
