--- src/lib/ui/ui.js
@@ -1,8 +1,0 @@
-// ui.js
-// UI helper functions
-
-// ===== PUBLIC API =====
-
-export function updateStatus(statusDiv, message) {
-  statusDiv.textContent = message;
-}
